requirements.txt
Flask
requests
python-dotenv
gunicorn
python-docx
psycopg2-binary
flask_cors
